provider: set defaults for all optional fields
Our proto3-basd API collapses missing (null) and zero ("", 0) values
to the zero value. Terraform doesn't like this. When a config omits an
optional attribute, it expects to see null in the state, not empty
string, 0, empty set, etc.

To fix this, we mark all optional properties as computed and set the
zero value as the default. Terraform will supply this value if not
explicitly set in the config, and thus be happy when that same zero
value appears in the state.

This commit also fixes the initialization of default values for sets.