context_question: remove excluded_blueprint_series_id field
Previously, the provider supported setting this exclusion relationship
on both the blueprint (excluded_context_question_series_ids) and on
the context question (excluded_blueprint_series_ids).

This is footgun, because it easy for the two sides to conflict. In
Terraform, each "piece" of config should be managed by a single
resource.

This PR standardizes on the blueprint resource for this exclusion.  We
believe that exclusions will generally be due to some unique aspect of
the specific blueprint, so managing the exclusions directly on it will
be the most convenient.